Remove the need of pa_module
Chrome supports all the components that are needed to transmit audio to
the Chromecast. This patch rewrites the Chrome app to create a
RTCPeerConnection and attach the PACC virtual source to it.
PulseAudio supports virtual sink-source pairs which makes the pa_module
project obsolete. This highly simplifies the way people can use this
project as the complicated toolchain setup is no longer necessary.

Also switches to the new Chromecast framework and introduces the cast
button.